+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 
 # 🚀Changelog
 -------------------------------------------------------------------------------------------------------------
-# 5.8.44(2026-02-01)
+# 5.8.44(2026-02-03)
 ### 🐣新特性
 * 【core   】      `NumberUtil.parseNumber`增加支持科学计数法（pr#4211@Github）
 * 【captcha】      `AbstractCaptcha`增加`setStroke`方法支持线条粗细（issue#IDJQ15@Gitee）
@@ -16,6 +16,7 @@
 * 【core       】  修复`TransMap.computeIfAbsent`mappingFunction处理不一致问题（issue#IDM6UR@Gitee）
 * 【core       】  修复`MultiResource`游标歧义问题（issue#IDNAOY@Gitee）
 * 【core       】  修复`BufferUtil`copy歧义问题（issue#IDN097@Gitee）
+* 【core       】  修复`JschSessionPool`回收导致的session未关闭问题（issue#4223@Github）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.43(2026-01-04)